Run YouCompleteMe only if its folder is found

+++ update.sh
@@ -382,17 +382,19 @@ if [[ NEED_VIM_PLUGIN_INSTALL -ne 0 ]]; then
     vim +PluginUpdate +qall
     echo $VIM_PLUGIN_HASH > $SCRIPTPATH/faaltu/.vim_plugin.hash
   fi
-  YCM_HASH=$(git -C $HOME/.vim/bundle/YouCompleteMe/ rev-parse @)
-  if [ ! -f $SCRIPTPATH/faaltu/.clang+llvm.hash ]; then
-    YCM_OLD_HASH="Nothing"
-  else
-    YCM_OLD_HASH=$(cat $SCRIPTPATH/faaltu/.clang+llvm.hash)
-  fi
-  if [ "$YCM_HASH" != "$YCM_OLD_HASH" ]; then
-    $HOME/.vim/bundle/YouCompleteMe/install.py --clang-completer --java-completer
-    echo $YCM_HASH > $SCRIPTPATH/faaltu/.clang+llvm.hash
-  else 
-    echo "YouCompleteMe is latest"
+  if [ -f $HOME/.vim/bundle/YouCompleteMe/install.py ]; then
+    YCM_HASH=$(git -C $HOME/.vim/bundle/YouCompleteMe/ rev-parse @)
+    if [ ! -f $SCRIPTPATH/faaltu/.clang+llvm.hash ]; then
+      YCM_OLD_HASH="Nothing"
+    else
+      YCM_OLD_HASH=$(cat $SCRIPTPATH/faaltu/.clang+llvm.hash)
+    fi
+    if [ "$YCM_HASH" != "$YCM_OLD_HASH" ]; then
+      $HOME/.vim/bundle/YouCompleteMe/install.py --clang-completer --java-completer
+      echo $YCM_HASH > $SCRIPTPATH/faaltu/.clang+llvm.hash
+    else 
+      echo "YouCompleteMe is latest"
+    fi
   fi
   wget --version > /dev/null
   if [ "$?" == 0 ]; then
